requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1

# PDF processing
pdfplumber>=0.10.3
pdf2image>=1.16.3
Pillow>=10.1.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0

# AI/ML
google-generativeai>=0.3.1
spacy>=3.7.2

# API framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing (optional for now)
pytest>=7.4.3
pytest-asyncio>=0.21.1

